+++ Changelog.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.9] - 2024-05-25
+
+### Summary
+
+- Design sqlmanager engine -- inprogress
+
+### Added
+
+- Create `sanbox` and `response.inspector` function in `reaction.js` at webnodejs engine which will internally handle error inside the trycatch from any components module. Completed on 2024-05-25
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+[1.0.8]: https://github.com/wkloh76/oricommjs/releases/tag/1.0.9
+
 ## [1.0.8] - 2024-05-14
 
 ### Summary